# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar categorias, transacciones, caja y logs. Documentación generada con swaggo.",
        "title": "ATM API",
        "contact": {
            "name": "Soporte",
            "email": "soporte@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/caja": {
            "get": {
                "description": "Si solo_caja=true, devuelve solo el saldo local. Si no, devuelve el saldo combinado con Odoo.",
                "produces": [
                    "application/json"
                ],
                "summary": "Obtener saldo en caja",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Solo saldo de caja local (sin Odoo)",
                        "name": "solo_caja",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Caja"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/categorias": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Listar categorias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Categoria"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Crear categoria",
                "parameters": [
                    {
                        "description": "Categoria",
                        "name": "categoria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CategoriaCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Obtener categoria por id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Actualizar categoria (parcial)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar (parcial)",
                        "name": "categoria",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.CategoriaUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Categoria"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Eliminar categoria",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/limpiar": {
            "post": {
                "description": "Elimina todas las transacciones, logs, caja, y categorías (deja la estructura vacía)",
                "produces": [
                    "application/json"
                ],
                "summary": "Eliminar todos los datos de la base de datos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/logs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Listar logs de transacciones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/notify/test": {
            "get": {
                "description": "Envía un mensaje de prueba usando las variables de entorno NOTIFY_*. Por defecto envía \"ping\".",
                "produces": [
                    "application/json"
                ],
                "summary": "Enviar mensaje de prueba de notificación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Texto a enviar",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/resumen": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Obtener resumen financiero (vista resumen_financiero)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Caja": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "saldo": {
                    "type": "number"
                },
                "ultima_actualizacion": {
                    "type": "string"
                }
            }
        },
...
        "models.Categoria": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
        "models.CategoriaCreateInput": {
            "type": "object",
            "required": [
                "nombre",
                "tipo"
            ],
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
        "models.CategoriaUpdateInput": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
...
    }
}